+++ aws/static-website/modules/cloudfront_logging/main.tf
@@ -1,0 +1,27 @@
+resource "aws_cloudwatch_log_delivery_source" "this" {
+  provider = aws.us_east_1
+
+  name         = "${var.name}-cloudfront-source"
+  log_type     = "ACCESS_LOGS"
+  resource_arn = var.aws_cloudfront_distribution_arn
+}
+
+resource "aws_cloudwatch_log_delivery_destination" "this" {
+  provider = aws.us_east_1
+
+  name = "${var.name}-cloudfront-destination"
+  delivery_destination_configuration {
+    destination_resource_arn = "${var.log_bucket_arn}/${var.log_bucket_prefix_path}"
+  }
+}
+
+resource "aws_cloudwatch_log_delivery" "this" {
+  provider = aws.us_east_1
+
+  delivery_source_name     = aws_cloudwatch_log_delivery_source.this.name
+  delivery_destination_arn = aws_cloudwatch_log_delivery_destination.this.arn
+
+  s3_delivery_configuration {
+    suffix_path = "/{DistributionId}/{yyyy}/{MM}/{dd}/{HH}"
+  }
+}