+++ aws/static-website/variables.tf
@@ -212,8 +212,21 @@ variable "versioning_status" {
 }
 
 # logging
-variable "log_bucket_domain_name" {
-  description = "Log bucket's domain name used to store access log from CloudFront."
+
+variable "enable_logging" {
+  description = "Whether to enable logging (V2)."
+  type        = bool
+  default     = false
+}
+
+variable "log_bucket_arn" {
+  description = "Log bucket's arn that used to store CloudFront logs."
+  type        = string
+  default     = null
+}
+
+variable "log_bucket_prefix_path" {
+  description = "Log bucket's prefix path to store CloudFront logs."
   type        = string
   default     = null
 }